agentlog: move http-debug into new file
To have a better overview - SURPRISE!
